サウンド再生処理修正　AudioSource追加してなかった

+++ project/Assets/Script/GameScene/TitleMainGameScene.cs
@@ -29,14 +29,14 @@ public class TitleMainGameScene : GameScene {
 	public override void init(){
 		Debug.Log("to Scene title");
 		initObjects (mList);
-		SoundController.PlaySoundBGM(GameSound.BgmList.TITLE);
+		SoundController.PlaySoundBGM((int)GameSound.BgmList.TITLE);
 	}
 	/*
 	 * 更新;
 	 */
 	public override void update(){
 		if (InputManager.isTouchObject("btn_start(Clone)")) {
-			SoundController.PlaySoundSE(GameSound.SeList.GAME_START);
+			SoundController.PlaySoundSE((int)GameSound.SeList.GAME_START);
 			SceneController.setChangeScene(SceneController.Scene.GAME_MAIN);
 		}
 	}
